refactor lib: clean directories + clean options + add theming capabilities + remove usage of ionicon (for arrows)
