+++ example/fonts/ionicons.ttf
