--- index.js
@@ -1,6 +1,0 @@
-import JsonView from './json-viewer';
-
-const install = (Vue) => {
-    Vue.component('JsonViewer', JsonView);
-}
-export default Object.assign(JsonView, {install});
