+++ lib/index.js
@@ -1,0 +1,7 @@
+import JsonView from './json-viewer'
+
+const install = Vue => {
+  Vue.component('JsonViewer', JsonView)
+}
+
+export default Object.assign(JsonView, {install})
